+++ http/http-common/src/main/java/ru/tinkoff/kora/http/common/body/StreamingHttpBodyOutput.java
@@ -21,7 +21,7 @@ public class StreamingHttpBodyOutput extends AtomicBoolean implements HttpBodyOu
     }
 
     @Override
-    public int contentLength() {
+    public long contentLength() {
         return this.contentLength;
     }
